Update `SlackChannelFlags`
Added a few flags per this document from the Slack APIs: https://api.slack.com/types/conversation

- `is_ext_shared`
- `is_pending_ext_shared`
- `is_read_only`

Also alphabetized flags so it's easier to compare with Slack's docs that are alphabetized.

Note that `is_user_deleted` isn't in the Slack docs for this response type, but it is on the im type here: https://api.slack.com/types/im

I wasn't sure if that is being used somehow so I left it alone.